+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Improve the Page Manager experience when dragging and dropping pages - the updates happen in background and the UI is not blocked anymore.
 * Allow scrolling while dragging a page in the Page Manager.
 * Change user's email field type to `email`.
+* Improve media manager experience after uploading images. No additional server requests are made, no broken UI on error.
 * Change reset password form button label to `Reset Password`.
 
 ### Fixes